Разбил код на модули, подключил точку входа в index.html
